вывод в окошке about сетевых адресов

+++ shared/data_types.h
@@ -58,9 +58,9 @@ public:
 struct NRecStats {
     QTime time;
     double avg_bitrate;
-    size_t size;
+    quint64 size;
 
-    NRecStats(QTime time=QTime(), double avg_bitrate=0., size_t size=0) {
+    NRecStats(QTime time=QTime(), double avg_bitrate=0., quint64 size=0) {
         this->time=time;
         this->avg_bitrate=avg_bitrate;
         this->size=size;
